rm circular dep
