+++ src/lib/swapper/swappers/OneInchSwapper/getMinMax/getMinMax.test.ts
@@ -1,6 +1,6 @@
 import { FOX, WETH } from '../../utils/test-data/assets'
 import type { OneInchSwapperDeps } from '../utils/types'
-import { getMinMax } from './helpers'
+import { getMinMax } from './getMinMax'
 
 jest.mock('../getUsdRate/getUsdRate', () => ({
   getUsdRate: () => {
